Initial checkin of some sample OSKI sources. I assume that some of the stuff in impls/ will change significantly, particularly due to the boot sequence. I'm not completely happy with some of the names, particularly Service and ServiceNotify. The simplest name was Controller, but that was just too general.
The basic outline is this:

To use an OSKI service, you instantiate and wire to a component
from services/.

Services generally refer to an Impl component, which is the completely
packaged up version of that service. A ServiceImpl component takes the
underlying ServiceC component (the actual service) and wires it for
initialization and the boot sequence. The ServiceC component wires an
underlying components (such as ServiceP, or implementations in
tos/system or tos/lib) as need be. The P was supposed to be Private, or
something, and isn't important; it will probably change once we decide
on a file naming convention. Broadcast shows a good example of this
breakdown.

One thing that has been separated is service start/stop and
notification that it has started/stopped (interfaces Service and
ServiceNotify, respectively). When you introduce things like send
queues, perfect split-phase synchronization isn't always necessary. I
think this approach is definitely unsuitable to lower level
components, but above the service boundary can simplify things a lot.


git-svn-id: http://tinyos-main.googlecode.com/svn@14 285f258d-8064-b571-fbc6-9656dc998e90