initial checkin of platform independent CC2420 radio stack uses the new Init, SplitControl interfaces with error_t. Also uses the new TOSMsg format with   TOSMsg.header,   TOSMsg.footer,   TOSMsg.metadata defined by the radio in the chips/radio directory (in our case, CC2420Const.h). The common fields of a TOSMsg are exported by the platform indpeendent RadioPacket interface from RadioC.  If the radio is a CSMA radio, it also must contain CSMARadioC (see the code for details).
This implementation is not guarenteed to compile or work, it is simply
a placeholder to see how TinyOS 2.x primitives will work with the CC2420
radio stack.  Once basic applications are working in TinyOS 2.x, this
"chip" will be debugged and the corresponding commit messages will indicate
when it is safe to use.


git-svn-id: http://tinyos-main.googlecode.com/svn@41 285f258d-8064-b571-fbc6-9656dc998e90