ATmega128 Timer (HPL Layer)
Added interrupt handlers and access functions for manipulating
the interrupt mask and flag registers:
    start (enable mask bit)
    stop  (disable mask bit)
    reset (clear flag bit)
    test  (read flag bit)
    check (read mask bit)

Most timer functionality can now be implemented in a HAL such that switching
from one timer to another (of like kind, i.e. 8 vs. 16 bit) can be done with
a wiring change.  Further abstraction of the various capture channels is best
done in the HAL.


git-svn-id: http://tinyos-main.googlecode.com/svn@43 285f258d-8064-b571-fbc6-9656dc998e90