MathOpsM.nc is a source file with basic operations for builtin integer types.
MathOpsM.perl.nc is a source pattern that produces MathOpsM.nc, because doing
that stuff by hand is pretty painful, escially when adding or changing
operations.

MakePerlnesc is the script that invokes perlncpp to produce MathOpsM.nc.

WidenAlarmM and WidenCounterM take AlarmBase and CounterBase interfaces and
widen then to a greater number of bits to the developer's specification.  I
itend these to be used with the Cast and eventual Resample interfaces so a
developer only need implement code for the native timers and their native
widths, using these generics to abstract up to Milli's and 32-bits and such.

CastAlarmM and CastCounterM take 32-bit AlarmBase and CounterBase interfaces
and re-exposes them as standard 32-bit Alarm and Counter interfaces.


git-svn-id: http://tinyos-main.googlecode.com/svn/branches/tinyos-2_0_devel-BRANCH@93 285f258d-8064-b571-fbc6-9656dc998e90