Executing timers can get tricky when you consider that an executing timer can change the state (start, stop, etc) of any other timer, including itself. This is the common case for periodic timers, which for all intents and purposes are one-shot timers that are preconfigured to re-arm themselves each time they fire.
This is gets complicated by the way the code manages the sense of "now" as it
processes the timers.  Before, the code considered t0 to be strictly in the
past from "now".  To be consistent in processing the timers, "now" has to be
anchored at the start of iterating over the current timer set.  So, that "now"
slips into the past as timers are processed.  If in the due course of
processing, a timer re-arms itself with respect to the true "now", then it is
effectively anchored in the future with respect to the currently running timer
execution loop.

This is a mess to work around unless the code uses the new assumption that t0
can be in the future.  Essentially, the difference of (now-t0) is a signed
value, and if negative, t0 is in the future.  This is fine, though the valid
ranges of (t0,dt) have changed.

  Before:
    now-MAX <= t0 <= now
    0 <= dt <= MAX

  Now:
    now-MAX/2 <= t0 <= now+MAX/2
    0 <= dt <= MAX/2

So, the maximum value of dt is halved, but t0 can be specified to start in the
future.  As a fun test, I ran Blink specifying that the timers should start 5
seconds after the node booted, which works.  Cool, I guess.

I've also removed the linked lists from the code.  Now each time the timers
are executed, all timers are checked to see if they're set.  This makes
smaller and more robust code.


git-svn-id: http://tinyos-main.googlecode.com/svn@206 285f258d-8064-b571-fbc6-9656dc998e90