Some broad sweeping changes to component naming in networking stacks. Specifically:
Chips do not have global component names. E.g., the CC1000 does
not have a CSMARadioPacketC component. Instead, a platform is
responsible for mapping the expected global names (e.g.,
ActiveMessageC) to the platform's chip names. This allows a
platform to have, for example, more than one radio, without
relying on -I path hackery to get the right one as the primary.

This naming change has implications for packet structures as well;
a platform maps the global names (e.g., RadioHeader) to the
chip names (e.g., CC1000Header).

Active messages are now a platform-specific implementation, as per
the telecon discussion two weeks ago. In the CC1000 case,
cc1000/SendReceive is responsible for sending, receiving, and
acks, while CC1000ActiveMessageM performs the dispatch.

TOS_LOCAL_ADDRESS has been removed, as it is not necessarily
a variable on all platforms (Joe's mention, without
generating an AM address from an underlying id) makes
this necessary. There is a "common" implementation in system,
which is just a component that stores an AM address. You can
still specify it at the command line or at boot time, etc. This
decomposition into a separate component gives platforms flexibility
and a common (but not omnipresent) global name, decide how to
implement it, while not introducing problematic
dependencies between different parts of a the networking stack.


git-svn-id: http://tinyos-main.googlecode.com/svn@423 285f258d-8064-b571-fbc6-9656dc998e90