You can boot motes in TOSSIM now, through the Python interface. Only the Null app works, though, as there's zero hardware support.
Finally figured out the right naming and levels of indirection between
C -> C++ -> Python. Also, SWIG doesn't like the standard TinyOS
convention of inttypes, as it doesn't know how to deal with them.
Later it might be useful to add typemaps for this purpose, but I'd
rather just get things working for now.


git-svn-id: http://tinyos-main.googlecode.com/svn@722 285f258d-8064-b571-fbc6-9656dc998e90