Interface updates so they're useful.
The one questionable choice here is my modification to Capture;
I removed the notion than on a capture event you can disable
further captures through a return value. This is a holdover semantics
from early pre-1.x code and is really an easy way to introduce bugs.
That being said, it could be an important primitive, I'm not sure.
It seems to me like calling disable() in the handler is better.
BUT, if it turns out for concurrency reasons that you need this
mechanism, then it should most certainly NOT be done with an
error_t, rather with a bool.


git-svn-id: http://tinyos-main.googlecode.com/svn/branches/tinyos-2_0_devel-BRANCH@733 285f258d-8064-b571-fbc6-9656dc998e90