Changed packets to respect TinyOS naming conventions and to follow Vlado's suggestions. types/message.h now defines message_t, and the file platform_message.h defines message_header_t, message_footer_t, and message_metadata_t. I've kept the old files around for now: the have a cpp warning to indicate that their use should be updated.
I tried to scour the repository as well as I could to update everything.
I am uncertain about intelmote2, because I was unable to get xscale-elf-gcc
to work properly (it didn't like my libm.a). But I'm able to compile for
telosb, micaz, mica2, and eyesIFXv2 (which has a few atomic warnings),
have tested micaz/telosb interoperatility, and the serial test application
works.


git-svn-id: http://tinyos-main.googlecode.com/svn@1169 285f258d-8064-b571-fbc6-9656dc998e90