Removed enum in Atm128Adc.h which was preventing compilation -- need to check with DGay on why it was there.
Updated atm128 SPI so it supports TEP 112 (Mcu Power management):
calls McuSleep when needed.

Updated simulation implementation of counter so that it does not
do an snprintf if debugging is disabled: otherwise, given that it's
in a timer interrupt emulation routine, there's a pretty big effect
on performance (e.g., a 10% slowdown). Instead, used sim_time_string()
function inline, which is cut out when debugging is turned off.


git-svn-id: http://tinyos-main.googlecode.com/svn@1698 285f258d-8064-b571-fbc6-9656dc998e90