I've updated the usart directory so that USART0 is properly shared. I have not yet updated USART1 in case we want to make any more modifications to the updated structure.
Starting from the bottom up, the arbiter and resource interfaces have
been removed from the lowest level HplMsp430Usart0C component, so it
is entirely stateless (except for l_br, l_mctl, etc). Instead,
Msp430Usart0C provides an arbitrated version of the HplUsart
interfaces. This shim provides arbitration and usart events are only
dispatched to current owners of the resource.

Above Msp430Usart0C sits the SPI components (and potentially I2C and
UART, but they are not implemented). Currently, the
Msp430Spi{NoDma,Dma}0P component takes a parameterized resource
interface for clients to arbitrate and uses a parameterized resource
interface to arbitrate the usart. SPI over DMA is now supported but it
is not wired in by default. A one-line change in Msp430Spi0C will
allow the SPI to use DMA.

By having the SPI component use a parameterized resource interface,
individual SPI requests are arbitrated with Msp430Usart0C's
arbiter. While this means less code space (no need for another
arbiter), it does mean that the SPI components cannot arbitrate with a
separate priority. Also, this structure requires calling enableSPI on
every time the resource is acquired, but there's a check in enableSPI
to quickly return if it's arleady in SPI mode. This brings up the
concern that SPI requests cannot be satisfied in batch. I'd argue that
latency is no concern since granting requests require a task post
anyways. Futhermore, from what I've seen, a bunch of SPI commands are
executed while a client holds the resource so much of the batching
occurs there anyways.

We can talk through what the right decision is, or maybe we should
just provide both implementations and let people choose. One nice
thing about this architecture is that both implementations are easy to
do.

I know that the eyes platform would like to have a different policy
for managing the USART by allowing the radio to have highest priority
access. To support this, it will have to shadow Msp430Usart0C.nc and
Msp430UsartShare0P.nc and wire in the appropriate arbiter and power
manager.

Some other minor changes:
- I've separated the events from HplMsp430Usart into a separate
interface called HplMsp430UsartInterrupts.
- Removed ArbiterInfo from Spi0C. I couldn't find anyone that actually
used this interface.
- Renamed Spi0C to Msp430Spi0C since it is msp430 specific.


git-svn-id: http://tinyos-main.googlecode.com/svn@1713 285f258d-8064-b571-fbc6-9656dc998e90