A bunch of bug fixes:
1) Sukun's send problem was due to ForwardingEngineP not freeing a send
queue entry when it sent a packet: the memory leak cause it to think that
its queue was full.

2) Found a bug in the interactions between the router and the link
estimator where the router could receive beacon packets it did not
originate. The current fix is a hack, but works. The estimator sends
its own beacons if it thinks it needs to; as they are the same AM type,
received LE beacons are signaled to the router. But as the router did
not put a payload into them, if it looks for one it reads garbage. Also,
the Router's calculation of payload size was incorrect: should be
sizeof(beacon_msg_t), not sizeof(beacon_msg_t*).

Test the tree formation and link estimation with simple topology. Seems
to be working OK. Nodes prefer links with more hops if the metric is
smaller. It's tough to determine how good the ETX measure is, but it seems
in the right ballpark.


git-svn-id: http://tinyos-main.googlecode.com/svn/branches/tinyos-2_0_devel-BRANCH@2045 285f258d-8064-b571-fbc6-9656dc998e90