Background: The current architecture of the usart HPL/HAL on the msp430 does not support multiple clients that require different configurations of the usart. It assumes that all SPI users want the same number of chars, the same clock polarity, etc.  The commands that we have for changing the clock source and the rate are actually not working reliably in multi-client setting, since according to TI these actions can not be made "on-the-fly" but only while the USART is in "reset" state. Resetting reverts all config registers to their default state. So putting the USART in a given mode and configuration have to be done in the same go. Consequently, if we want to enable multiple users we need to apply "complete" reconfiguring of the device between each client, using the ResourceConfigure interface. This brings the overhead of storing and writing ~ couple of words on each arbitration handover, but I don't see any alternatives. For I2C the same applies for the part of the configuration that can only be done while I2C module is disabled.
Current status:
I have introduced a new header file (Msp430Usart.h), with three new structures that contain all the necessary configuration information for a SPI, UART or I2C session (a collection of flags and settings from several different registers). This are mostly the settings that HAVE to be specified during SWRST=1 or I2CEN=0. For the flags/registers that can be RW dynamically on character by character basis (like multiprocessor UART where one can specify if the next byte is address or data; or like in I2CSTP etc.) we need dedicated commands in the interface.

In addition to these "configuration" structures, there are 1:1 mirrored structures of the different registers where these flags come from. For lack of time, I decided to skip the I2C registers. On configuration, the registers are populated using the settings from the configuration structures ultimately specified by the end-clients using protocol specific interfaces like Msp430SpiConfigure and Msp430UartConfigure. When the client does not provide a configuration, the protocol abstraction for Spi and Uart should provide sane default settings.

Next steps:
- Do the same for I2C
- Change the configs to operate using pointers to the config structures
- Fix the serial stack so that it does not assume that it is the only user of the UART
- Find a solution that ensures that the default command handler is executed when exporting the interface using generic configurations and the client does not wire to it


git-svn-id: http://tinyos-main.googlecode.com/svn/branches/tos-2-msp430-usart-configure-candidate@2198 285f258d-8064-b571-fbc6-9656dc998e90