Add alternative LED abstractions.
This branch introduces new abstractions for LEDs.  The existing Leds
interface supports only three LEDs, and encodes the LED number within the
command name.  OSIAN's primary development platform is the SuRF board, which
has five LEDs.  The branch introduces two new interfaces:

 - Led presents an interface to a single LED, with the same on/off/toggle
   options as Leds, adding a set command that takes a boolean as argument.

 - MultiLed presents an interface to a collection of LEDs, providing the
   same integer-valued get/set commands as Leds, adding individual
   on/off/toggle/set commands where the affected LED is a command parameter.

These interfaces are supported by the LedC component, which presents a
parameterized Led interface, individual Led interfaces for each numbered LED
on the platform, individual Led interfaces for LEDs of specific colors, and
the MultiLed interface.  The number of LEDs supported and the color of each
one are defined by a PlatformLed.h header and implemented in a PlatformLedC
component.

Among other things, this allows applications to (for example) always use the
red LED as a specific indicator without requiring the application to
identify the target platform and call the specific Leds command associated
with that LED.  The standard LedC component supports up to eight LEDs, and
colors Green, Red, White, Yellow, Orange, and Blue.

Tests/demonstrations of the interfaces have been added in
apps/tests/TestLed.  Some of these depend on osian/serialprintf.

Demonstration support for these interfaces has been added within telosa,
with a component that simply defines the interfaces in terms of the
underlying LedsC component.  Other platforms can probably just clone these
files.  (I don't know what the telosa actually had, but that's where the LED
definitions for telosb live and since I've got a couple of those I put it in
that platform.)

git-svn-id: http://tinyos-main.googlecode.com/svn@5258 285f258d-8064-b571-fbc6-9656dc998e90