DMA-driven UartStream for msp430.
This isn't really from OSIAN; it's taken from Steve's RPL repository and
provides a DMA-based UartStream implementation on MSP430.  The HdlcFramingC
component for the PPP daemon does a lot of heavy processing on a
per-character basis during reception, and without this characters tend to
get dropped at baud rates above 9600.  A similar problem is observed on the
SuRF hardware, though in that case the problem is alleviated by increasing
the clock speed.  Reworking the MSP430 infrastructure to support DMA on the
UART generally would be a good thing.

git-svn-id: http://tinyos-main.googlecode.com/svn@5262 285f258d-8064-b571-fbc6-9656dc998e90