Merging external cortex development branch into cortex-devel.
Most conflicts were in the move of the rf2xx directory over to tos/lib/rfxlink. We currently still need the rf2xx directory for opal. However, I am not sure if my merge deleted some lines, or if I made mistakes in keeping the right versions. We need to verify this.

git-svn-id: http://tinyos-main.googlecode.com/svn@5282 285f258d-8064-b571-fbc6-9656dc998e90