Revert "DMA-driven UartStream for msp430."
This reverts commit 59fbd1b82c12cd0ed07b0886d46215deaead8399.

The approach taken in that commit supports DMA for RX only, not TX;
hard-codes a dependency on DMA channel 2, and changes the behavior of the
UartStream buffer-based receive infrastructure to produce a stream of
receiveDone events rather than the standard one event per receive command.
An alternative approach will be used to solve the problems this patch was
intended to address.

git-svn-id: http://tinyos-main.googlecode.com/svn/branches/osian-squashed-ppp@5439 285f258d-8064-b571-fbc6-9656dc998e90