Provide a new UART interface better suited to the needs of HdlcFramingC.
The HDLC framing infrastructure needs the ability to send blocks of
characters, and to receive characters one-by-one as soon as they arrive.
The processing done on each received character is fairly complex.

The UartStream interface's send command and receiveByte interface
technically meet these needs.  However, both are async operations.  At high
serial data rates and when serving as a bridge for a high-data-rate radio
interface, interrupt-driven reception causes dropped packets.

DMA-based reception can work around this, but DMA reception requires a
receive buffer, and the behavior of UartStream is such that if a buffer is
installed the receivedByte event required for HdlcFraming is not signalled.
Instead, timely delivery of characters requires snooping the in-progress DMA
operation using an Alarm or similar functionality, as well as coping with
the unexpected mix of reception completion events.

Rather than attempt to provide a generic DMA-based UART solution that
maintains the existing behavior of the UartStream interface, we provide a
new HdlcUart interface that does exactly what's needed.  A default
implementation based on the PlatformSerialC UartStream implementation is
provided, with a buffer that is used to allow received character processing
outside of async context.  For platforms where this is inadequate, or where
extremely high serial rates are desired, a DMA-based implementation of
HdlcUart can be provided that does not impact the standard behavior of
UartStream or PlatformSerialC.

git-svn-id: http://tinyos-main.googlecode.com/svn/branches/osian-squashed-ppp@5440 285f258d-8064-b571-fbc6-9656dc998e90