Support TI headers and newer MSP430 chips.
This branch adds two features, combined because they interfere with each
other as both involve changes to how presence/absence of port modules are
detected.

 - It adds support for the MSP430 chips that have resistor control.  There
   was some discussion of this approach in
   https://www.millennium.berkeley.edu/pipermail/tinyos-devel/2010-January/004221.html.
   It does extend the existing HplMsp430GeneralIO with two new commands, but
   I believe the impact on existing code is minimal or non-existent.

 - It adds support for mspgcc4 builds that use the Texas Instruments header
   files.  These change the names of certain functional presence
   preprocessor directives.

Forewarning: The current approach for port interfaces will not work with the
upcoming uniarch changes to mspgcc, where the address of the port registers
will no longer be available as a constant.  A completely different approach
will be required once uniarch is available.

git-svn-id: http://tinyos-main.googlecode.com/svn@5455 285f258d-8064-b571-fbc6-9656dc998e90